chore: 폴더 구조 수정, 필요없는 파일 삭제, 파일명 수정
